Replace "whitelist" option with "allowlist"
The term "whitelist" is a metaphorical term used to say "allow these
items". While it is extremely common in technical works, it can
inadvertantly promote the trope of "white is good, black is bad". For
more details on the term and why "allowlist" is a suitable, more
technically accurate replacement, see:
https://tools.ietf.org/id/draft-knodel-terminology-01.html#rfc.section.1.2

BREAKING CHANGE: The "whitelist" option has been removed and replaced
with "allowlist"